requirements.txt
app
fastapi
flask
uvicorn
pydrive
pydantic
pandas
openpyxl
python-dotenv
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
slack_sdk
textblob==0.17.1
APScheduler
pytz

